Finished Dragonfly.
Changed generator to spawn from 2 possible sets. set A is an obstacle or danger, set B is a
collectable with a chance of spawning a danger at the same spot. The chance of spawning
danger with collectable increases with difficulty curve.
Messed with lighting to help make flower look less bright.