Preliminary work on Main Menu, and Save Data.
Added in Save Load system, and have plans of attaching achievements to it.
With Achievements comes unlockables, and with unlockables comes an equip menu.

Added in Gloria Font, and started working on general UI Theme.
On Main Menu I plan on using custom texture buttons.
On Submenues I plan on using the General UI Theme called Bee_Theme.

Added in Hit VFX for player.
Removed the Intro Screen from the HUD, and began work on migrating it's function to the Main
Menu.
Plan on adding a Results screen as well, with the goal of emphasizing the players
accomplishments and showcasing what they achieved and unlocked.